requirements.txt
# Core
torch>=2.1.0
numpy>=1.24.0

# Speech Recognition
openai-whisper

# Language Model
llama-cpp-python>=0.2.0

# Text-to-Speech
pyttsx3>=2.90

# Audio
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0